chore: split backend module to a few files
