+++ core/src/commands/logout.ts
@@ -12,8 +12,8 @@ import { printHeader } from "../logger/util.js"
 import { dedent } from "../util/string.js"
 import { clearAuthToken, getStoredAuthToken } from "../cloud/legacy/auth.js"
 import { getCloudDomain } from "../cloud/util.js"
-import { gardenBackendFactory } from "../cloud/backend.js"
 import { findProjectConfigOrPrintInstructions } from "./helpers.js"
+import { gardenBackendFactory } from "../cloud/backend-factory.js"
 
 type Opts = {}
